Tests/LibWeb/Text/expected/setInterval.txt

setInterval completed count=0
setInterval completed count=1
setInterval completed count=2
